mainnet: make the workdir persistent
This change mounts the complete working directory on the host.
The config sub-directory is populated on the first run only.
Config can now be edited and loaded with a `docker-compose restart`.

The working directory now lives in `/data` and is defined
as persistent VOLUME in the Dockerfile.
The `node` binary is now copied to /usr/local/bin/node.
This way the executable and persistent data are better seperated.

The keygenerator can only be run after config is mounted.
Hence, it is now included in the node-obs image
and ran in the entrypoint script, only on first run.
Keygenerator is removed from the start_stack.sh script.

Note that proxy did not have mountpoints defined in docker-compose
nor run-proxy.sh.
References to the proxy/config mounts in start_stack.sh
and README.md have been cleaned up.